# ToDoListAPI/Entities/Role.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ToDoListAPI.Entities
{
    public class Role
    {
        public int Id { get; set; }
        public string Name { get; set; }
    }
}